+++ packages/signers/signer/src/index.ts
@@ -1,1 +1,2 @@
+export type { PolkadotSigner } from "@polkadot-api/polkadot-signer"
 export * from "./from-raw-signer"