+++ packages/codegen/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Generate metadata types instead of checksums [#561](https://github.com/polkadot-api/polkadot-api/pull/561)
 
+## 0.6.2 - 2024-07-11
+
+### Fixed
+
+- Update dependencies
+
 ## 0.6.1 - 2024-07-03
 
 ### Fixed