+++ packages/signers/polkadot-signer/CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Unreleased
 
+## 0.1.0 - 2024-07-11
+
+### Changed
+
+- Add API to sign raw-data and rename `sign` to `signTx`
+
 ## 0.0.2 - 2024-07-03
+
+### Fixed
 
 - Update dependencies